fastApi/requirements.txt
mangum
fastapi
pydantic
psycopg2-binary
python-multipart
uvicorn
boto3